refactor: 모달 컴포넌트 문서화 및 구조 정리 및 사용 예제 추가
