+++ src/components/modal/modals/ReviewRegisterModal/ReviewRegisterModal.types.ts
@@ -1,0 +1,18 @@
+export type ReviewTasteKey = 'body' | 'tannin' | 'sweet' | 'acid';
+
+export type ReviewRegisterValue = {
+  rating: number;
+  content: string;
+  taste: Record<ReviewTasteKey, number>; // 0~100
+  aromas: string[];
+};
+
+export type ReviewRegisterModalProps = {
+  isOpen: boolean;
+  onClose: () => void;
+
+  wineName: string;
+  wineImageUrl?: string;
+
+  onSubmit: (value: ReviewRegisterValue) => Promise<void> | void;
+};
